I cooked, implemented layers for CNN, RNN, DNN and some util layers.

+++ utils/optimizer.py
@@ -45,8 +45,8 @@ class AdamOptimizer(Optimizer):
 
     def _initialize_param_state(self, param: np.ndarray, name: str, param_id: int):
         if 'm' not in self.param_states[param_id]:
-            self.param_states[param_id]['m'] = np.zeros_like(param, dtype=np.float32)
-            self.param_states[param_id]['v'] = np.zeros_like(param, dtype=np.float32)
+            self.param_states[param_id]['m'] = np.zeros_like(param, dtype=np.float64)
+            self.param_states[param_id]['v'] = np.zeros_like(param, dtype=np.float64)
 
     def _update_single_param(self, param: np.ndarray, grad: np.ndarray, lr: float):
         param_id = id(param)
@@ -83,7 +83,7 @@ class SGD(Optimizer):
 
     def _initialize_param_state(self, param: np.ndarray, name: str, param_id: int):
         if 'velocity' not in self.param_states[param_id]:
-            self.param_states[param_id]['velocity'] = np.zeros_like(param, dtype=np.float32)
+            self.param_states[param_id]['velocity'] = np.zeros_like(param, dtype=np.float64)
 
     def _update_single_param(self, param: np.ndarray, grad: np.ndarray, lr: float):
         param_id = id(param)
